+++ packages/core/src/connectors/coinbaseWallet.ts
@@ -38,6 +38,8 @@ export class CoinbaseWalletConnector extends Connector<
       provider.on('chainChanged', this.onChainChanged)
       provider.on('disconnect', this.onDisconnect)
 
+      this.emit('connecting')
+
       const accounts = await provider.enable()
       const account = getAddress(accounts[0])
       const id = await this.getChainId()
@@ -102,8 +104,10 @@ export class CoinbaseWalletConnector extends Connector<
   }
 
   async getSigner() {
-    const provider = this.getProvider()
-    const account = await this.getAccount()
+    const [provider, account] = await Promise.all([
+      this.getProvider(),
+      this.getAccount(),
+    ])
     return new providers.Web3Provider(
       <ExternalProvider>(<unknown>provider),
     ).getSigner(account)