+++ packages/react/CHANGELOG.md
@@ -1,5 +1,41 @@
 # wagmi
 
+## 0.3.0-next.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wagmi/core@0.2.0-next.6
+
+## 0.3.0-next.7
+
+### Minor Changes
+
+- don't persist account data when `autoConnect` is falsy
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wagmi/core@0.2.0-next.5
+
+## 0.3.0-next.6
+
+### Minor Changes
+
+- fix `useContractWrite` mutation fn arguments
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wagmi/core@0.2.0-next.4
+
+## 0.3.0-next.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @wagmi/core@0.2.0-next.3
+
 ## 0.3.0-next.4
 
 ### Minor Changes